agents/requirements.txt
uvicorn
python-dotenv
google-adk
fastapi
pydantic
firebase-admin
google-genai